Add Azure Storage Queue Event Bus and enhance options
Added AddAzureStorageQueueEventBus method in ServiceExtensions.cs to register Azure Storage Queue Event Bus with XML documentation and optional configuration. Modified StorageQueueFactory to pass StorageQueueOptions to StorageQueueProcessor. Enhanced StorageQueueOptions with XML documentation and new MessageProcessing class. Updated StorageQueueProcessor to use StorageQueueOptions for message processing behavior.

+++ src/Arcturus.EventBus.AzureStorageQueue/StorageQueueProcessor.cs
@@ -7,13 +7,15 @@ public sealed class StorageQueueProcessor : IProcessor
 {
     private readonly StorageQueueConnection _connection;
     private readonly string _queueName;
+    private readonly StorageQueueOptions _options;
 
     public event Func<IEventMessage, OnProcessEventArgs?, Task>? OnProcessAsync;
 
-    internal StorageQueueProcessor(IConnection connection, string? queueName = null)
+    internal StorageQueueProcessor(IConnection connection, StorageQueueOptions storageQueueOptions, string? queueName = null)
     {
         _connection = (StorageQueueConnection)connection;
         _queueName = queueName ?? "default_queue";
+        _options = storageQueueOptions;
     }
 
     public async Task WaitForEvents(CancellationToken cancellationToken = default)
@@ -26,7 +28,8 @@ public sealed class StorageQueueProcessor : IProcessor
 
         while (!cancellationToken.IsCancellationRequested)
         {
-            var receiveMessages = await queueClient.ReceiveMessagesAsync(1, null, cancellationToken);
+            var receiveMessages = await queueClient.ReceiveMessagesAsync(
+                1, _options.MessageProcessing.VisibilityTimeout, cancellationToken);
 
             if (receiveMessages is not null && receiveMessages.Value.Length > 0)
             {
@@ -37,19 +40,27 @@ public sealed class StorageQueueProcessor : IProcessor
 
                     if (OnProcessAsync is not null)
                     {
+                        if (_options.MessageProcessing.DeleteMessageBeforeProcessing.GetValueOrDefault(true)) 
+                        { 
+                            // delete message before it resurface
+                            await queueClient.DeleteMessageAsync(
+                                receiveMessage.MessageId, receiveMessage.PopReceipt, cancellationToken);
+                        }
+
+                        // raise the request
+                        // TODO: try/catch on exception move to dead letter
                         await OnProcessAsync.Invoke(
                             @event
                             , new OnProcessEventArgs(
                                 receiveMessage.MessageId
                                 , null
                                 , cancellationToken));
-                        await queueClient.DeleteMessageAsync(
-                            receiveMessage.MessageId, receiveMessage.PopReceipt, cancellationToken);
                     }
                 }
             }
 
-            cancellationToken.WaitHandle.WaitOne(100);
+            cancellationToken.WaitHandle.WaitOne(
+                _options.MessageProcessing.MessageIntervalMilliseconds ?? 100);
         }
     }
 }
